+++ src/components/ModalInfo/ModalInfo.module.scss
