+++ src/components/ModalInfo/ModalInfo.tsx
